Modules and packages
In this example, we define a package called "package" that contains two modules, "module1" and "module2". We then import the entire "package" in another script called "main.py" and call the "add" and "subtract" functions from the "module1" and "module2" modules, respectively, with the arguments 5 and 3. The output of the script is 8 and 2, respectively.